+++ package.json
@@ -4,7 +4,10 @@
   "description": "Practice app for course by Andrew Mead",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "serve": "live-server public/",
+    "babel": "babel src/app.js --out-file public/scripts/app.js --presets=env,react --watch",
+    "build": "webpack --watch"
   },
   "repository": {
     "type": "git",
@@ -19,5 +22,11 @@
   "dependencies": {
     "babel-preset-env": "^1.5.2",
     "babel-preset-react": "^6.24.1"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "live-server": "^1.2.1",
+    "webpack": "^5.28.0",
+    "webpack-cli": "^4.6.0"
   }
 }
