+++ public/index.html
@@ -9,6 +9,6 @@
     <div id="app"></div>
     <script crossorigin src="https://unpkg.com/react@16/umd/react.development.js"></script>
     <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
-    <script src="/scripts/app.js"></script>
+    <script src="bundle.js"></script>
 </body>
 </html>
