+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path')
+
+module.exports = {
+    entry: './src/app.js',
+    output: {
+        path: path.join(__dirname, 'public'),
+        filename: 'bundle.js'
+    }
+};
